# openapi.yaml
openapi: 3.0.3
info:
  title: coldemailmpv API
  version: 0.1.0
  description: |
    MVP backend for cold email SaaS.

    Current scope includes:
    - JWT auth
    - workspaces and members
    - sender accounts with SMTP and IMAP config
    - prospects and CSV import
    - one-off sends and email events
    - one-step campaigns
    - suppression list
    - public unsubscribe endpoint
servers:
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        message:
          type: string
        path:
          type: string
        errors:
          nullable: true
          oneOf:
            - type: object
              additionalProperties:
                type: string
            - type: array
              items:
                type: string
            - type: string
      required: [timestamp, status, message, path]
...
    CreateSenderRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        smtpHost:
          type: string
        smtpPort:
          type: integer
          minimum: 1
          maximum: 65535
        smtpUsername:
          type: string
        smtpPassword:
          type: string
        fromName:
          type: string
          nullable: true
        startTls:
          type: boolean
          default: true
        ssl:
          type: boolean
          default: false
        imapHost:
          type: string
          nullable: true
        imapPort:
          type: integer
          nullable: true
          minimum: 1
          maximum: 65535
        imapUsername:
          type: string
          nullable: true
        imapPassword:
          type: string
          nullable: true
        imapSsl:
          type: boolean
          default: true
      required: [email, smtpHost, smtpPort, smtpUsername, smtpPassword]
    SenderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        smtpHost:
          type: string
        smtpPort:
          type: integer
        smtpUsername:
          type: string
        fromName:
          type: string
          nullable: true
        startTls:
          type: boolean
        ssl:
          type: boolean
        imapHost:
          type: string
          nullable: true
        imapPort:
          type: integer
          nullable: true
        imapUsername:
          type: string
          nullable: true
        imapSsl:
          type: boolean
        lastImapSyncAt:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
        createdAt:
          type: string
          format: date-time
      required: [id, email, smtpHost, smtpPort, smtpUsername, startTls, ssl, imapSsl, status, createdAt]
    SenderTestResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
    SendTestEmailRequest:
      type: object
      properties:
        to:
          type: string
          format: email
        subject:
          type: string
          maxLength: 255
        text:
          type: string
          maxLength: 10000
      required: [to, subject, text]
    SendTestEmailResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        messageId:
          type: string
          nullable: true
        sentAt:
          type: string
          format: date-time
          nullable: true
      required: [success, message]
...
    UnsubscribeRequest:
      type: object
      properties:
        token:
          type: string
      required: [token]
    UnsubscribeResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        email:
          type: string
          format: email
      required: [success, message, email]
security:
  - bearerAuth: []
paths:
...
  /api/senders:
    get:
      tags: [Senders]
      summary: List sender accounts
      responses:
        '200':
          description: Sender accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SenderResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Senders]
      summary: Create sender account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSenderRequest'
      responses:
        '200':
          description: Sender created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SenderResponse'
        '400':
          description: Validation or business error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/senders/{senderId}/test:
    post:
      tags: [Senders]
      summary: Test SMTP connection for sender
      parameters:
        - name: senderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: SMTP test result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SenderTestResponse'
        '400':
          description: SMTP error or business error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/senders/{senderId}/test-imap:
    post:
      tags: [Senders]
      summary: Test IMAP connection for sender
      parameters:
        - name: senderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: IMAP test result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SenderTestResponse'
        '400':
          description: IMAP error or business error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/senders/{senderId}/send-test:
    post:
      tags: [Senders]
      summary: Send a test email using sender account
      parameters:
        - name: senderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendTestEmailRequest'
      responses:
        '200':
          description: Test email send result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendTestEmailResponse'
        '400':
          description: Validation, SMTP or business error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/public/unsubscribe:
    get:
      tags: [Public]
      summary: Unsubscribe by token via query param
      security: []
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Unsubscribe success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnsubscribeResponse'
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Public]
      summary: Unsubscribe by token via JSON body
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnsubscribeRequest'
      responses:
        '200':
          description: Unsubscribe success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnsubscribeResponse'
        '400':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'