# src/main/java/ru/sakhapov/emailwarmup/api/dto/UnsubscribeRequest.java
package ru.sakhapov.emailwarmup.api.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class UnsubscribeRequest {

    @NotBlank
    private String token;
}